Add files via upload
My work on entrepreneurship in geospatial education was undertaken to obtain the Senior University Teacher Qualification (SUTQ). In this work, I analyzed a vast array of surveys spanning six years, examining feedback from 883 students. Based on this analysis and a literature review, I designed a course that addressed students' needs for skills in understanding and applying entrepreneurial concepts in geospatial education.